Add Proper Long Press & Haptics, Reorganize Files/Classes/Structs, Remove Unnecessary VC & VM
