+++ Calculator/Calculator/Calculator/Utility/Data Types/Generator.swift
